Zero out clippy and flutter analyze; gate CI on both
Rust: name the pending-registration tuple type in auth.rs, give the
scenario-test module a non-shadowing name, convert build_progression_message's
13 positional arguments to a ProgressionMessage params struct (five call sites
were passing a dozen values positionally), and group the schplanner test
helper's prescription into SingleGroupSpec.

Flutter: wrap single-statement ifs in app_theme, resolve AuthService before the
async gap in passkeys_screen, exclude generated lib/gen/** from analysis, and
mark the sound generator as a CLI tool where printing is the point.

CI now runs clippy with -D warnings and flutter analyze --fatal-infos, so both
stay at zero.

Also adds Dart tests for the pure logic modules: plate calculator (greedy
loading, remainder bounds, sweep invariants) and weight units
(conversion, snapping idempotence, formatting). 35 Flutter tests, up from 6.

+++ examples/api_invariant_fuzz.rs
@@ -167,7 +167,7 @@ async fn run_user(
 
         let group_count = rng.gen_range(1..4);
         let groups: Vec<ExerciseGroup> = (0..group_count)
-            .map(|i| random_group(&mut rng, i as i32))
+            .map(|i| random_group(&mut rng, i))
             .collect();
 
         let started = match wk
